# Put uuids in the vae and eav index