--- a/server/resources/migrations/73_vae_uuids.down.sql
+++ b/server/resources/migrations/73_vae_uuids.down.sql
@@ -1,0 +1,5 @@
+drop function public.json_uuid_to_uuid(v jsonb);
+
+drop index vae_uuid_index;
+
+drop index eav_uuid_index;