--- a/server/resources/migrations/73_vae_uuids.up.sql
+++ b/server/resources/migrations/73_vae_uuids.up.sql
@@ -1,0 +1,15 @@
+create or replace function public.json_uuid_to_uuid(v jsonb) returns uuid
+  language sql immutable
+  as $$
+   select (v->>0)::uuid
+  end;
+$$;
+
+-- Create this index concurrently before running the migration
+create index if not exists vae_uuid_index
+  on triples(app_id, public.json_uuid_to_uuid(value), attr_id, entity_id)
+  where vae;
+
+create unique index if not exists eav_uuid_index
+  on triples(app_id, entity_id, attr_id, public.json_uuid_to_uuid(value))
+  where eav;