We only put refs in the vae and eav indexes so we know that the value will be a uuid. This PR stores those uuids as actual uuids in the index itself. 

This makes the indexes take up less space (they're about 30% smaller).

The big change is that now we can use `vae` instead of `eav` when we do a join on a ref, which saves us doing an expensive json->uuid conversion in the join filter. Here's a simple example:

```clojure
{:books {:where {:category "western" :author :eid/elmer-kelton}}}
```

That generates two ctes:

```
with m_0 as (
  select entity_id as m_0_eid from triples t0 where attr_id = :category-eid and value = "western"
),
m_1 as (
  select m_0.*, entity_id as m_1_eid from triples t1 where value = :eid/elmer-kelton and m_0.m_0_eid = json_to_uuid(value)
)
```

If we use the `eav` index for m_1, the `v` is at the end of the index. Postgres does a nested loop and filters the join on `entity_id = json_to_uuid(value)`. It looks something like this:

```
    ->  Nested Loop
          Join Filter: (t0.entity_id = ((t1.value ->> 0))::uuid)
          Rows Removed by Join Filter: N
          ->  Index Scan using ave_index on triples t0
                Index Cond: (attr_id = :category-eid) AND (value = '"western"'::jsonb)
          ->  Index Scan using eav_uuid_index on triples t1
                Index Cond: ((entity_id = :eid) AND (attr_id = :author-eid))
```

If we use the `vae` index for m_1, then the `v` is at the start of the index and postgres will use that in the index cond instead of the less efficient join filter. It looks something like this:

```
    ->  Nested Loop 
          ->  Index Scan using ave_index on triples t0 
                Index Cond: (attr_id = :category-eid) AND (value = '"western"'::jsonb)
          ->  Index Scan using vae_uuid_index on triples t1
                Index Cond: ((((value ->> 0))::uuid = t0.entity_id) AND (attr_id = :author-eid) AND (entity_id =:eid))
```

`((value ->> 0))::uuid` is already in the index, so we never have to do any json->uuid conversion at all in the query itself. We avoid the expensive join filter, which in my backtesting makes queries much faster.

We also don't have to carry around the `match_x_x_value_uuid` any more. Instead of `value_uuid`, we track `is_ref_value` so that we know to convert it from a string to a uuid in `sql-row->triple`.

<details>
<summary>Results from the backtest:</summary>

I'm not super concerned about the first two queries that got worse because they're not slower if I enable pg hints.

| :old-explain | :new-explain | :improvement |
|-------------|--------------|--------------|
|          513 |         2075 |        -1562 |
|          676 |         1578 |         -901 |
|          393 |          574 |         -181 |
|          489 |          595 |         -105 |
|          286 |          375 |          -89 |
|            3 |           86 |          -83 |
|            6 |           87 |          -80 |
|            4 |           83 |          -78 |
|            5 |           84 |          -78 |
|            5 |           84 |          -78 |
|            6 |           84 |          -78 |
|          341 |          418 |          -76 |
|            5 |           82 |          -76 |
|            5 |           79 |          -74 |
|          306 |          379 |          -73 |
|            5 |           70 |          -65 |
|           99 |          164 |          -64 |
|          290 |          351 |          -61 |
|          110 |          171 |          -60 |
|          280 |          340 |          -60 |
|          117 |          175 |          -58 |
|          134 |          192 |          -58 |
|          117 |          173 |          -56 |
|           80 |          135 |          -55 |
|            4 |           59 |          -55 |
|          144 |          198 |          -54 |
|            5 |           58 |          -53 |
|            4 |           58 |          -53 |
|          110 |          163 |          -53 |
|          125 |          177 |          -52 |
|            4 |           56 |          -51 |
|          115 |          166 |          -51 |
|          131 |          181 |          -49 |
|          238 |          288 |          -49 |
|          156 |          205 |          -49 |
|          139 |          188 |          -49 |
|            4 |           53 |          -49 |
|           87 |          136 |          -48 |
|          445 |          494 |          -48 |
|            4 |           53 |          -48 |
|          457 |          502 |          -45 |
|          136 |          181 |          -44 |
|          110 |          153 |          -43 |
|          187 |          230 |          -43 |
|          134 |          177 |          -42 |
|          100 |          143 |          -42 |
|          112 |          154 |          -41 |
|           58 |           99 |          -41 |
|           22 |           63 |          -41 |
|          274 |          316 |          -41 |
|          119 |          160 |          -41 |
|           22 |           63 |          -41 |
|           87 |          128 |          -40 |
|          134 |          173 |          -38 |
|          120 |          158 |          -38 |
|           20 |           58 |          -38 |
|          379 |          417 |          -38 |
|          129 |          167 |          -38 |
|          103 |          139 |          -36 |
|          112 |          148 |          -35 |
|          125 |          159 |          -34 |
|           18 |           52 |          -33 |
|          113 |          147 |          -33 |
|          103 |          136 |          -33 |
|          109 |          141 |          -32 |
|           82 |          113 |          -31 |
|          170 |          201 |          -30 |
|           90 |          121 |          -30 |
|          196 |          226 |          -30 |
|          113 |          141 |          -28 |
|           73 |          100 |          -27 |
|           96 |          123 |          -27 |
|          157 |          184 |          -26 |
|           79 |          105 |          -26 |
|          205 |          231 |          -25 |
|            7 |           33 |          -25 |
|           29 |           54 |          -24 |
|          277 |          302 |          -24 |
|           82 |          106 |          -24 |
|           74 |           97 |          -23 |
|          278 |          302 |          -23 |
|          155 |          178 |          -22 |
|          350 |          371 |          -21 |
|            8 |           29 |          -20 |
|          299 |          320 |          -20 |
|           17 |           38 |          -20 |
|           76 |           97 |          -20 |
|           72 |           93 |          -20 |
|           10 |           30 |          -19 |
|          149 |          167 |          -18 |
|           20 |           38 |          -18 |
|           66 |           84 |          -18 |
|           19 |           37 |          -18 |
|          246 |          264 |          -17 |
|           42 |           59 |          -17 |
|           96 |          113 |          -17 |
|           70 |           88 |          -17 |
|           82 |          100 |          -17 |
|           36 |           53 |          -16 |
|           29 |           45 |          -15 |
|            6 |           21 |          -15 |
|            4 |           19 |          -15 |
|            3 |           18 |          -15 |
|            9 |           24 |          -15 |
|           38 |           53 |          -15 |
|          120 |          135 |          -15 |
|           61 |           76 |          -15 |
|            3 |           18 |          -14 |
|           54 |           69 |          -14 |
|           98 |          113 |          -14 |
|           93 |          108 |          -14 |
|          113 |          127 |          -14 |
|            3 |           17 |          -14 |
|            3 |           17 |          -13 |
|            7 |           20 |          -12 |
|           42 |           55 |          -12 |
|          315 |          328 |          -12 |
|          220 |          232 |          -11 |
|            4 |           15 |          -11 |
|           10 |           21 |          -11 |
|           37 |           47 |          -10 |
|           39 |           49 |          -10 |
|           28 |           38 |           -9 |
|           44 |           54 |           -9 |
|           21 |           31 |           -9 |
|           43 |           53 |           -9 |
|           10 |           20 |           -9 |
|           33 |           42 |           -9 |
|           26 |           35 |           -9 |
|           30 |           39 |           -8 |
|           72 |           81 |           -8 |
|           47 |           55 |           -8 |
|           50 |           58 |           -8 |
|           53 |           61 |           -7 |
|            4 |           12 |           -7 |
|           44 |           51 |           -7 |
|           52 |           59 |           -6 |
|           37 |           44 |           -6 |
|           24 |           31 |           -6 |
|           56 |           63 |           -6 |
|           97 |          103 |           -6 |
|           57 |           64 |           -6 |
|           80 |           87 |           -6 |
|           48 |           54 |           -5 |
|          128 |          134 |           -5 |
|           51 |           57 |           -5 |
|            3 |            8 |           -5 |
|           40 |           45 |           -5 |
|            3 |            8 |           -4 |
|          120 |          125 |           -4 |
|           28 |           33 |           -4 |
|           50 |           55 |           -4 |
|          129 |          133 |           -4 |
|            2 |            6 |           -4 |
|           20 |           24 |           -4 |
|           21 |           25 |           -4 |
|           20 |           24 |           -4 |
|           12 |           16 |           -3 |
|           40 |           44 |           -3 |
|           47 |           51 |           -3 |
|          174 |          178 |           -3 |
|           19 |           22 |           -3 |
|            9 |           12 |           -3 |
|           17 |           20 |           -3 |
|           27 |           31 |           -3 |
|           17 |           21 |           -3 |
|            9 |           12 |           -3 |
|           18 |           21 |           -3 |
|           12 |           15 |           -3 |
|           41 |           44 |           -3 |
|          132 |          135 |           -3 |
|           19 |           22 |           -2 |
|           30 |           33 |           -2 |
|            9 |           12 |           -2 |
|           36 |           38 |           -2 |
|           18 |           21 |           -2 |
|            9 |           11 |           -2 |
|           13 |           15 |           -2 |
|            8 |           10 |           -2 |
|            4 |            6 |           -1 |
|           11 |           13 |           -1 |
|           44 |           45 |           -1 |
|           29 |           31 |           -1 |
|           10 |           12 |           -1 |
|           14 |           15 |           -1 |
|            2 |            4 |           -1 |
|           99 |          101 |           -1 |
|           15 |           16 |           -1 |
|           10 |           12 |           -1 |
|           40 |           41 |           -1 |
|           14 |           15 |           -1 |
|           27 |           28 |           -1 |
|           73 |           74 |           -1 |
|            6 |            8 |           -1 |
|            8 |            9 |           -1 |
|           49 |           50 |           -1 |
|           15 |           16 |           -1 |
|           20 |           21 |           -1 |
|           13 |           14 |           -1 |
|           12 |           13 |           -1 |
|            6 |            7 |           -1 |
|            6 |            7 |           -1 |
|            4 |            5 |           -1 |
|           22 |           23 |           -1 |
|           33 |           34 |            0 |
|            6 |            7 |            0 |
|           14 |           15 |            0 |
|            7 |            8 |            0 |
|            1 |            2 |            0 |
|           16 |           16 |            0 |
|           57 |           58 |            0 |
|            6 |            6 |            0 |
|           14 |           15 |            0 |
|           11 |           12 |            0 |
|            7 |            8 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|            2 |            2 |            0 |
|            4 |            5 |            0 |
|           18 |           19 |            0 |
|           16 |           16 |            0 |
|            1 |            2 |            0 |
|            4 |            5 |            0 |
|           19 |           20 |            0 |
|           10 |           11 |            0 |
|            2 |            2 |            0 |
|            2 |            2 |            0 |
|            4 |            5 |            0 |
|           21 |           22 |            0 |
|            4 |            4 |            0 |
|            3 |            3 |            0 |
|            2 |            3 |            0 |
|            8 |            8 |            0 |
|            8 |            8 |            0 |
|            1 |            1 |            0 |
|           30 |           31 |            0 |
|            6 |            7 |            0 |
|           23 |           23 |            0 |
|            7 |            8 |            0 |
|            4 |            4 |            0 |
|            7 |            8 |            0 |
|            6 |            7 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|            4 |            5 |            0 |
|            8 |            8 |            0 |
|           18 |           18 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|            1 |            1 |            0 |
|           29 |           29 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|           10 |           10 |            0 |
|            4 |            4 |            0 |
|            2 |            2 |            0 |
|            3 |            4 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|           12 |           12 |            0 |
|            1 |            1 |            0 |
|            1 |            2 |            0 |
|            1 |            1 |            0 |
|           13 |           13 |            0 |
|            8 |            8 |            0 |
|            1 |            2 |            0 |
|            3 |            4 |            0 |
|            0 |            1 |            0 |
|            0 |            0 |            0 |
|            1 |            1 |            0 |
|           15 |           15 |            0 |
|           28 |           28 |            0 |
|            1 |            1 |            0 |
|            9 |           10 |            0 |
|           18 |           18 |            0 |
|            2 |            2 |            0 |
|            2 |            2 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|            1 |            1 |            0 |
|        30000 |        30000 |            0 |
|        30000 |        30000 |            0 |
|        30000 |        30000 |            0 |
|            1 |            1 |            0 |
|            5 |            5 |            0 |
|            2 |            2 |            0 |
|           56 |           56 |            0 |
|            5 |            5 |            0 |
|            2 |            2 |            0 |
|            1 |            1 |            0 |
|            1 |            1 |            0 |
|            9 |            9 |            0 |
|            1 |            1 |            0 |
|            2 |            2 |            0 |
|            7 |            7 |            0 |
|            2 |            2 |            0 |
|           27 |           27 |            0 |
|            3 |            2 |            0 |
|            1 |            1 |            0 |
|            7 |            7 |            0 |
|            3 |            3 |            0 |
|            5 |            5 |            0 |
|            5 |            5 |            0 |
|           14 |           14 |            0 |
|            6 |            6 |            0 |
|           10 |           10 |            0 |
|            2 |            1 |            0 |
|           34 |           34 |            0 |
|           23 |           23 |            0 |
|            2 |            2 |            0 |
|            4 |            3 |            0 |
|            7 |            7 |            0 |
|            1 |            1 |            0 |
|           17 |           16 |            0 |
|            2 |            1 |            0 |
|            4 |            3 |            0 |
|            5 |            4 |            0 |
|            6 |            6 |            0 |
|            1 |            1 |            0 |
|           16 |           15 |            0 |
|            3 |            2 |            0 |
|           13 |           13 |            0 |
|            2 |            1 |            0 |
|            2 |            2 |            0 |
|            1 |            1 |            0 |
|           10 |           10 |            0 |
|            1 |            1 |            0 |
|            9 |            9 |            0 |
|            4 |            3 |            0 |
|            5 |            5 |            0 |
|            2 |            1 |            0 |
|            7 |            7 |            0 |
|            5 |            5 |            0 |
|            2 |            2 |            0 |
|            3 |            3 |            0 |
|            7 |            7 |            0 |
|            3 |            3 |            0 |
|            8 |            7 |            0 |
|            1 |            1 |            0 |
|            7 |            7 |            0 |
|            3 |            2 |            0 |
|            8 |            8 |            0 |
|            7 |            6 |            0 |
|            6 |            5 |            0 |
|           22 |           21 |            0 |
|            7 |            7 |            0 |
|           14 |           14 |            0 |
|           14 |           13 |            0 |
|            7 |            7 |            0 |
|            3 |            2 |            0 |
|            7 |            7 |            0 |
|            4 |            3 |            0 |
|           53 |           52 |            0 |
|            6 |            5 |            0 |
|            9 |            8 |            0 |
|           13 |           12 |            0 |
|            2 |            1 |            0 |
…